once more

+++ cmd/session_cruncher/session_cruncher.go
@@ -447,7 +447,7 @@ func TopSessionsThread() {
 					message.Keys = append(message.Keys, fmt.Sprintf("%016x_accelerated_percent", buyers[i]))
 					buyerAcceleratedPercent := 0.0
 					if buyerStats.totalSessions[i] > 0 {
-						buyerAcceleratedPercent = float64(buyerStats.nextSessions[i]/buyerStats.totalSessions[i]) * 100.0
+						buyerAcceleratedPercent = float64(buyerStats.nextSessions[i])/float64(buyerStats.totalSessions[i]) * 100.0
 					}
 					message.Values = append(message.Values, float64(buyerStats.totalSessions[i]))
 					message.Values = append(message.Values, float64(buyerStats.nextSessions[i]))
